requirements.txt
google-generativeai==0.7.2
pillow==10.4.0
openai==1.37.1
Flask==3.0.3
python-dotenv==1.0.1